Refactoring
Script Migration to Python: To enhance user interactivity and streamline the overall user experience, we have reconfigured the original RepBox bash scripts into subprocesses within a Python framework. This modification allows for greater ease of use, as users can now execute each portion of the pipeline through a set of commands via a single Python script that calls each command as a subprocess. Furthermore, we have ensured that the individual parameters utilized in the original RepBox scripts are included in each command.

Included Dependencies: We have proactively addressed the challenge of acquiring certain dependencies for our package, considering the potential difficulties users may face due to variable websites and changes in hosting over time. As a solution, we have incorporated the required packages, such as HelitronScanner and SINE_Scan, into an archive file. This file will automatically decompress and install these specific dependencies for seamless integration within the package.

Setup Instructions & Scripts: To improve clarity during the setup process and installation of dependencies, we have renamed the installation script from install.sh to setup_repbox.sh. This change aims to facilitate users in understanding and executing the setup and installation steps for RepBox. Furthermore, the setup script has been thoroughly updated to align with the modifications made to the pipeline. As part of this update, we have also revised the installation instructions to accurately reflect the revised implementation of the pipeline script. Users can now benefit from clear and concise instructions that precisely correspond to the recent changes implemented in the package.